+++ app/screens/home/recent_mentions/components/channel_info/index.ts
@@ -1,0 +1,26 @@
+// Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
+// See LICENSE.txt for license information.
+
+import withObservables from '@nozbe/with-observables';
+import {switchMap, of as of$} from 'rxjs';
+
+import ChannelInfo from './channel_info';
+
+import type PostModel from '@typings/database/models/servers/post';
+import type TeamModel from '@typings/database/models/servers/team';
+
+const enhance = withObservables(['post'], ({post}: {post: PostModel}) => {
+    const channel = post.channel.observe();
+
+    return {
+        channelName: channel.pipe(
+            switchMap((chan) => of$(chan.displayName)),
+        ),
+        teamName: channel.pipe(
+            switchMap((chan) => chan.team || of$(null)),
+            switchMap((team: TeamModel|null) => of$(team?.displayName || null)),
+        ),
+    };
+});
+
+export default enhance(ChannelInfo);
