+++ app/screens/home/recent_mentions/recent_mentions.tsx
@@ -1,38 +1,26 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
-import {Q} from '@nozbe/watermelondb';
-import {withDatabase} from '@nozbe/watermelondb/DatabaseProvider';
-import withObservables from '@nozbe/with-observables';
 import {useIsFocused, useRoute} from '@react-navigation/native';
-import compose from 'lodash/fp/compose';
 import React, {useCallback, useState, useEffect, useMemo} from 'react';
 import {useIntl} from 'react-intl';
 import {StyleSheet, View, ActivityIndicator, FlatList} from 'react-native';
 import Animated, {useAnimatedStyle, withTiming} from 'react-native-reanimated';
 import {SafeAreaView, Edge} from 'react-native-safe-area-context';
-import {of as of$} from 'rxjs';
-import {switchMap} from 'rxjs/operators';
-
 import {getRecentMentions} from '@actions/remote/search';
 import DateSeparator from '@app/components/post_list/date_separator';
-import {MM_TABLES, SYSTEM_IDENTIFIERS} from '@app/constants/database';
-import {SystemModel, UserModel} from '@app/database/models/server';
+import {UserModel} from '@app/database/models/server';
 import {getDateForDateLine, isDateLine, selectOrderedPosts} from '@app/utils/post_list';
-import {getTimezone} from '@app/utils/user';
 import NavigationHeader from '@components/navigation_header';
-import {useServerUrl} from '@context/server_url';
+import {useServerUrl} from '@context/server';
 import {useTheme} from '@context/theme';
 import {useCollapsibleHeader} from '@hooks/header';
 
 import Mention from './components/mention';
 
-import type {WithDatabaseArgs} from '@typings/database/database';
 import type PostModel from '@typings/database/models/servers/post';
 
 const AnimatedFlatList = Animated.createAnimatedComponent(FlatList);
-
-const {USER, SYSTEM, POST} = MM_TABLES.SERVER;
 
 const EDGES: Edge[] = ['bottom', 'left', 'right'];
 
@@ -41,7 +29,6 @@ type Props = {
     currentUser: UserModel;
     isTimezoneEnabled: boolean;
     mentions: PostModel[];
-    theme: Theme;
 }
 
 const styles = StyleSheet.create({
@@ -171,36 +158,4 @@ const RecentMentionsScreen = ({mentions, currentUser, currentTimezone, isTimezon
     );
 };
 
-const enhance = withObservables([], ({database}: WithDatabaseArgs) => {
-    const currentUser = database.get<SystemModel>(SYSTEM).findAndObserve(SYSTEM_IDENTIFIERS.CURRENT_USER_ID).pipe(
-        switchMap((currentUserId) => database.get<UserModel>(USER).findAndObserve(currentUserId.value)),
-    );
-
-    return {
-        mentions: database.get<SystemModel>(SYSTEM).query(
-            Q.where('id', SYSTEM_IDENTIFIERS.RECENT_MENTIONS),
-            Q.take(1),
-        ).observeWithColumns(['value']).pipe(
-            switchMap((rows) => {
-                if (!rows.length || !rows[0].value.length) {
-                    return of$([]);
-                }
-                const row = rows[0];
-                return database.get(POST).query(
-                    Q.where('id', Q.oneOf(row.value)),
-                    Q.sortBy('create_at', Q.asc),
-                ).observe();
-            }),
-        ),
-        currentUser,
-        currentTimezone: currentUser.pipe((switchMap((user) => of$(getTimezone(user.timezone))))),
-        isTimezoneEnabled: database.get<SystemModel>(SYSTEM).findAndObserve(SYSTEM_IDENTIFIERS.CONFIG).pipe(
-            switchMap((config) => of$(config.value.ExperimentalTimezone === 'true')),
-        ),
-    };
-});
-
-export default compose(
-    withDatabase,
-    enhance,
-)(RecentMentionsScreen);
+export default RecentMentionsScreen;
